update to issue CS-2267

+++ campsite/src/classes/cache/TemplateCacheHandler_DB.php
@@ -17,13 +17,11 @@ class TemplateCacheHandler_DB extends TemplateCacheHandler
 
     /**
      * Clears template cache storage.
-     * @param $p_campsiteVector
-     * @return boolean
      */
-    public function clean()
+    public function clean($tpl_file = null)
     {
-        $content = $smrty_obj = null;
-        self::handler('clean', $smarty_obj, $cache_content);
+        $cache_content = $smrty_obj = null;
+        self::handler('clean', $smarty_obj, $cache_content, $tpl_file);
     }
 
     /**
@@ -35,15 +33,16 @@ class TemplateCacheHandler_DB extends TemplateCacheHandler
         $queryStr = 'DELETE FROM Cache WHERE ' . self::vectorToWhereString($campsiteVector);
         $g_ado_db->Execute($queryStr);
 
-        if ($campsiteVector['article']) {
+        if ($campsiteVector['language'] && $campsiteVector['publication']) {
             $whereStr = "language = {$campsiteVector['language']} AND ";
             $whereStr .= "publication = {$campsiteVector['publication']} AND ";
-            $whereStr .= "issue >= {$campsiteVector['issue']} AND ";
-            $queryStr = 'DELETE FROM Cache WHERE ' . $whereStr . "section = {$campsiteVector['section']} AND ";
-            $queryStr .= "article IS NULL";
-            $g_ado_db->Execute($queryStr);
-        }
-        if ($campsiteVector['issue']) {
+            $whereStr .= $campsiteVector['issue'] ? "issue >= {$campsiteVector['issue']} AND "
+            :'issue IS NULL AND ';
+            if ($campsiteVector['section']) {
+                $queryStr = 'DELETE FROM Cache WHERE ' . $whereStr . "section = {$campsiteVector['section']} AND ";
+                $queryStr .= "article IS NULL";
+                $g_ado_db->Execute($queryStr);
+            }
             $queryStr = 'DELETE FROM Cache WHERE ' . $whereStr . "section IS NULL AND article IS NULL";
             $g_ado_db->Execute($queryStr);
         }
@@ -72,26 +71,26 @@ class TemplateCacheHandler_DB extends TemplateCacheHandler
 
             switch ($action) {
             case 'read':
-                $whereStr = self::vectorToWhereString($campsiteVector);
-                $whereStr .= " AND template = '$tpl_file'";
+                if ($campsiteVector['language'] && $campsiteVector['publication']) {
+                    $whereStr = self::vectorToWhereString($campsiteVector);
+                    $whereStr .= " AND template = '$tpl_file'";
 
-                $queryStr = 'SELECT expired, content FROM Cache WHERE ' . $whereStr;
-                $result = $g_ado_db->GetRow($queryStr);
-                if ($result) {
-                    if ($result['expired'] > time()) {
-                        $cacheExists = true;
-                        $cache_content = $result['content'];
-                        $result = true;
+                    $queryStr = 'SELECT expired, content FROM Cache WHERE ' . $whereStr;
+                    $result = $g_ado_db->GetRow($queryStr);
+                    if ($result) {
+                        if ($result['expired'] > time()) {
+                            $cacheExists = true;
+                            $cache_content = $result['content'];
+                            $result = true;
+                        } else {
+                            // clear expired cache
+                            $queryStr = 'DELETE FROM Cache WHERE expired <= '. time();
+                            $g_ado_db->Execute($queryStr);
+                            $cacheExists = false;
+                        }
                     } else {
-                        // clear expired cache
-                        $queryStr = 'DELETE FROM Cache WHERE expired <= '. time();
-                        $g_ado_db->Execute($queryStr);
                         $cacheExists = false;
-                        $return = false;
                     }
-                } else {
-                    $cacheExists = false;
-                    $return = false;
                 }
                 break;
 
@@ -99,19 +98,22 @@ class TemplateCacheHandler_DB extends TemplateCacheHandler
                 // in case template changing should delete old cached templates
                 if ($cacheExists) {
                     $queryStr = 'DELETE FROM Cache WHERE template = ' . "'$tpl_file'";
-                    $g_ado_db->GetOne($queryStr);
+                    $g_ado_db->Execute($queryStr);
                 }
 
-                // insert new cached template
-                $queryStr = 'INSERT IGNORE INTO Cache ';
-                $queryStr .= '(' . implode(',', array_keys($campsiteVector));
-                $queryStr .=  ',template,expired,content) VALUES (';
-                foreach ($campsiteVector as $key => $value) {
-                    $queryStr .= !isset($value) ? 'NULL,' : $value . ',';
+                if ($campsiteVector['language'] && $campsiteVector['publication']) {
+
+                    // insert new cached template
+                    $queryStr = 'INSERT IGNORE INTO Cache ';
+                    $queryStr .= '(' . implode(',', array_keys($campsiteVector));
+                    $queryStr .=  ',template,expired,content) VALUES (';
+                    foreach ($campsiteVector as $key => $value) {
+                        $queryStr .= !isset($value) ? 'NULL,' : $value . ',';
+                    }
+                    $queryStr .= "'$tpl_file',$exp_time,'" . addslashes($cache_content) . "')";
+                    $g_ado_db->Execute($queryStr);
+                    $return = $g_ado_db->Affected_Rows() > 0;
                 }
-                $queryStr .= "'$tpl_file',$exp_time,'" . addslashes($cache_content) . "')";
-                $g_ado_db->Execute($queryStr);
-                $return = $g_ado_db->Affected_Rows() > 0;
                 break;
 
             case 'clean':
